Delete product functionality refined and alerts and confirmation added

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
+import Swal from 'sweetalert2';
 
 import {useDispatch} from 'react-redux'
 import {deleteProductAction} from '../actions/deleteProductAction'
@@ -8,9 +9,20 @@ const Product = ({product}) => {
     const dispatch = useDispatch();
 
     const confirmDeleteProduct = id => {
-
-
-        dispatch(deleteProductAction(id))
+        Swal.fire({
+            title: 'Are you sure?',
+            text: 'You won\'t be able to revert this!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, don\'t delete it!'
+        }).then((result) => {
+            if(result.value) {
+                dispatch(deleteProductAction(id))
+            }
+        })
     }
     
     return (
